added class component and trip-point extends from it
